divide the base line code from dubbo codec in interface
